backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==42.0.2
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
idna==3.6
jwcrypto==1.5.4
Markdown==3.5.1
MarkupSafe==2.1.5
oauthlib==3.2.2
psycopg==3.1.18
pycparser==2.22
PyJWT==2.8.0
pytz==2024.2
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
